fix for mcmcheck

+++ python/tools/mcmCheck.sh
@@ -97,6 +97,7 @@ then
   echo "   helpflag= $HELP"
   echo "   verbose= $VERBOSE"
   echo "   pause= $PAUSE"
+  echo "   weightsProvidedFlag= $WEIGHTSPROVIDED"
   echo " "
 fi
 
@@ -104,7 +105,7 @@ if [ "$HELP" == "true" ]
 then
   echo "mcmCheck compares inference result from framework against HW result using blob compiled with mvNCCompile -cpp"
   echo " "
-  echo "Usage: source ./mcmCheck.sh -n <prototxt_file> -w <caffemodel_file> -i <image_file> -b <blob_file> -r <result_file> -e <expected_file> -v -h -p"
+  echo "Usage: source ./mcmCheck.sh -n <prototxt_file> -w <caffemodel_file> -i <image_file> -b <blob_file> -r <result_file> -e <expected_file> -software -v -h -p"
   echo " "
   echo "    -n is needed for compilation and generation of expected reference results"
   echo "    -w is optional, if you do not supply weights then caffe generated weights will be used"
@@ -131,7 +132,8 @@ then
   return 2
 fi
 
-#----------------
+#-------------------------------------------------------------------------------------------------------------------------------
+
 if [ "$COMPILE" == "true" ] && [ "$WEIGHTSPROVIDED" == "false" ]
 then
   rm -f cpp.blob
@@ -140,25 +142,29 @@ then
   #note: this produces a Fathom_expected.npy file during compilation using an random array of data passed through the caffe model
   $MDK_HOME/projects/Fathom/src2/mvNCCompile.py $NETWORK --cpp    
 fi
-#---------------- compile blob from prototxt
+
+#---------------- compile software only blob from prototxt
+
 if [ "$COMPILE" == "true" ] && [ "$DISABLEHARDWARE" == "true" ] && [ "$WEIGHTSPROVIDED" == "true" ]
 then
   rm -f cpp.blob
   echo "compiling for software"
   echo "using weights provided"
-#  ./mvNCCompile.py $NETWORK --new-parser -w $WEIGHTS --cpp
   $MDK_HOME/projects/Fathom/src2/mvNCCompile.py $NETWORK --new-parser -w $WEIGHTS --cpp 
 fi
+
 #----------------
+
 if [ "$DISABLEHARDWARE" == "false" ] && [ "$COMPILE" == "true" ] && [ "$WEIGHTSPROVIDED" == "false" ]
 then
   rm -f cpp.blob
   echo "compiling for harware"
   echo "weights not provided - generating weights"
-#  ./mvNCCompile.py $NETWORK --new-parser --cpp
   $MDK_HOME/projects/Fathom/src2/mvNCCompile.py $NETWORK --new-parser --cpp --ma2480
 fi
+
 #----------------
+
 if [ "$DISABLEHARDWARE" == "false" ] && [ "$COMPILE" == "true" ] && [ "$WEIGHTSPROVIDED" == "true" ]
 then
   rm -f cpp.blob
@@ -169,25 +175,40 @@ then
 fi
 
 #---------------- generate reference/expected .npy outout with provided weights and provided image for inference
-if [ "$GENREFERENCE" == "true" ] && [ "$WEIGHTSPROVIDED" == "true" ]
+
+if [ "$GENREFERENCE" == "true" ] && [ "$WEIGHTSPROVIDED" == "true" ] && [ ! -z "$IMAGE" ]
 then
   echo "executing mcmGenRef.py with provided weights"
   rm -f Fathom_expected.npy # delete the results file generated during compilation, inference on provided image will be saved to Fathom_expected.npy file now
   python3 $MCM_HOME/python/tools/mcmGenRef.py --network $NETWORK --weights $WEIGHTS --image $IMAGE
+elif [ "$GENREFERENCE" == "true" ]
+then
+      echo "ERROR: You must provide an image for to generate an expected output from Caffe - exiting"
+      exit  
 fi
 #---------------- generate reference/expected .npy outout without provided weights and provided image for inference
-if [ "$GENREFERENCE" == "true" ] && [ "$WEIGHTSPROVIDED" == "false" ]
+
+if [ "$GENREFERENCE" == "true" ] && [ "$WEIGHTSPROVIDED" == "false" ]  && [ ! -z "$IMAGE" ]
 then
   echo "executing mcmGenRef.py without provided weights"
   rm -f Fathom_expected.npy
-  python3 $MCM_HOME/python/tools/mcmGenRef.py --network $NETWORK --image $IMAGE 
+  python3 $MCM_HOME/python/tools/mcmGenRef.py --network $NETWORK --image $IMAGE
+elif [ "$GENREFERENCE" == "true" ] && [ "$WEIGHTSPROVIDED" == "false" ]
+then
+      echo "ERROR: You must provide an image for to generate an expected output from Caffe, also note you have not provided weights - exiting"
+      exit  
 fi
+
 #---------------- run blob on HW
-if [ "$RUNHW" == "true" ]
+
+if [ "$RUNHW" == "true" ] && [ ! -z "$IMAGE" ]
 then
   echo "running 1st blob"
   rm -f $RESULT.npy 
-  python3 $MCM_HOME/python/tools/mcmRunHW.py --blob $BLOB --image $IMAGE --result $RESULT  
+  python3 $MCM_HOME/python/tools/mcmRunHW.py --blob $BLOB --image $IMAGE --result $RESULT 
+else
+      echo "ERROR: You must provide an image for inference on hardware - exiting"
+      exit  
 fi
 if [ "$COMPAREBLOBS" == "true" ]
 then
@@ -200,6 +221,7 @@ then
   echo "running 2nd blob"
   python3 $MCM_HOME/python/tools/mcmRunHW.py --blob $BLOB2 --image $IMAGE --result $RESULT
 fi
+
 #---------------- compare output to reference
 echo "comparing results"
 python3 $MCM_HOME/python/tools/mcmCheckRef.py --reference $EXPECTED --result $RESULT.npy
